+++ CHANGELOG.md
@@ -2,6 +2,9 @@
   * [ExUnit] Use ANSI escape codes in CLI output
   * [ExUnit] Include suite run time on CLI results
 
+* bug fix
+  * [Path] Fix a bug on `Path.expand` when expanding `~`
+
 # v0.8.1 (2013-02-17)
 
 * enhancements